+++ src/d-table/DTable.vue
@@ -342,7 +342,7 @@
       >
         <d-pagination
           :total-pages="totalPages"
-          :current-page="currentPage"
+          :current-page="internalCurrentPage"
           :current-page-siblings="currentPageSiblings"
           @page-changed="handlePageChange"
           :smart-color="smartColor"
@@ -411,7 +411,6 @@ import {
   onMounted,
   watch,
   onUnmounted,
-  watchEffect,
 } from "vue";
 import { computePosition, flip, shift, offset } from "@floating-ui/dom";
 import { sort } from "./utils/sort";
@@ -434,6 +433,7 @@ const emit = defineEmits([
   "sort",
   "export",
   "download-csv",
+  "async-table-update",
 ]);
 const isExpanded = ref(false);
 
@@ -442,6 +442,8 @@ const currentTable = ref(null);
 const viewportShrunkToMobile = ref(false);
 
 const expandedData = ref(null);
+
+const internalCurrentPage = ref(props.currentPage);
 
 const { exportCsv } = useCsvExport(props.generatedCsvName);
 
@@ -672,9 +674,16 @@ const handlePageChange = (currentPage) => {
   if (!props.asyncPagination) {
     scopedCurrentPage.value = currentPage;
   }
+  internalCurrentPage.value = currentPage;
   emit("page-updated", currentPage);
-  if (searchValue.value) {
-    emit("search", searchValue.value, props.currentPage);
+  // if (searchValue.value) {
+  //   emit("search", searchValue.value, currentPage);
+  // }
+  if (props.asyncPagination && props.asyncSearch) {
+    emit("async-table-update", {
+      page: currentPage,
+      search: searchValue.value,
+    });
   }
 };
 
@@ -737,19 +746,26 @@ watch(renderedColumns, (newVal, oldVal) => {
   }
 });
 
-watchEffect(() => {
-  if (props.asyncSearch) {
-    if (searchValue.value) {
-      emit("search", searchValue.value, props.currentPage);
-    }
+watch(
+  () => props.currentPage,
+  () => {
+    internalCurrentPage.value = props.currentPage;
   }
-});
+);
 
 watch(searchValue, () => {
-  if (!props.asyncPagination) {
-    scopedCurrentPage.value = 1;
+  scopedCurrentPage.value = 1;
+
+  if (props.asyncSearch) {
+    emit("search", searchValue.value, 1);
   }
-  emit("page-updated", 1);
+  internalCurrentPage.value = 1;
+  if (props.asyncSearch && props.asyncPagination) {
+    emit("async-table-update", {
+      page: 1,
+      search: searchValue.value,
+    });
+  }
 });
 
 const validateBackground = (background, index) => {