+++ storybook-static/assets/preview-a4f3775f.js
@@ -1,2 +1,0 @@
-import{_ as e}from"./iframe-caa3ebaa.js";import"../sb-preview/runtime.js";var a={docs:{renderer:async()=>{let{DocsRenderer:r}=await e(()=>import("./DocsRenderer-3PUGWF3O-65263b96.js"),["./DocsRenderer-3PUGWF3O-65263b96.js","./chunk-S4VUQJ4A-9e9d2fb1.js","./iframe-caa3ebaa.js","./index-d8ade8ea.js","./_commonjsHelpers-042e6b4d.js","./index-d475d2ea.js","./index-d37d4223.js","./index-6ba087f2.js","./floating-ui.dom-c3a31ea6.js","./index-356e4a49.js"],import.meta.url);return new r}}};export{a as parameters};
-//# sourceMappingURL=preview-a4f3775f.js.map